refact: rename 'compose' components and hide vars
