+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/filter/internal/FacetListStateImpl.kt
@@ -3,20 +3,24 @@ package com.algolia.instantsearch.compose.filter.internal
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import com.algolia.instantsearch.compose.filter.FacetListCompose
+import com.algolia.instantsearch.compose.filter.FacetListState
 import com.algolia.instantsearch.core.Callback
 import com.algolia.instantsearch.core.selectable.list.SelectableItem
 import com.algolia.search.model.search.Facet
 
 /**
- * [FacetListCompose] implementation.
+ * [FacetListState] implementation.
  *
  * @param facets state holding selectable facets list value
  */
-internal class FacetListComposeImpl(
+internal class FacetListStateImpl(
     facets: List<SelectableItem<Facet>>
-) : FacetListCompose {
+) : FacetListState {
 
     override var facets by mutableStateOf(facets)
     override var onSelection: Callback<Facet>? = null
+
+    override fun setItems(items: List<SelectableItem<Facet>>) {
+        this.facets = items
+    }
 }
