+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/searchbox/SearchQuery.kt
@@ -1,13 +1,11 @@
 package com.algolia.instantsearch.compose.searchbox
 
-import androidx.compose.runtime.Stable
 import com.algolia.instantsearch.compose.searchbox.internal.SearchBoxComposeImpl
 import com.algolia.instantsearch.core.searchbox.SearchBoxView
 
 /**
  * Search box query component for compose.
  */
-@Stable
 public interface SearchQuery : SearchBoxView {
 
     /**
@@ -21,15 +19,7 @@ public interface SearchQuery : SearchBoxView {
      * @param query received query input
      * @param isSubmit true if the value is a submit, otherwise false
      */
-    public fun onValueChange(query: String, isSubmit: Boolean) {
-        val onQuery = if (isSubmit) onQuerySubmitted else onQueryChanged
-        onQuery?.invoke(query)
-    }
-
-    override fun setText(text: String?, submitQuery: Boolean) {
-        query = text ?: ""
-        if (submitQuery) onQuerySubmitted?.invoke(text)
-    }
+    public fun onValueChange(query: String, isSubmit: Boolean)
 }
 
 /**
