+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/searchbox/internal/SearchBoxComposeImpl.kt
@@ -16,4 +16,20 @@ internal class SearchBoxComposeImpl(
     override var query by mutableStateOf(query)
     override var onQueryChanged: Callback<String?>? = null
     override var onQuerySubmitted: Callback<String?>? = null
+
+    /**
+     * the callback to be triggered on text update
+     *
+     * @param query received query input
+     * @param isSubmit true if the value is a submit, otherwise false
+     */
+    override fun onValueChange(query: String, isSubmit: Boolean) {
+        val onQuery = if (isSubmit) onQuerySubmitted else onQueryChanged
+        onQuery?.invoke(query)
+    }
+
+    override fun setText(text: String?, submitQuery: Boolean) {
+        query = text ?: ""
+        if (submitQuery) onQuerySubmitted?.invoke(text)
+    }
 }
