+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/stats/internal/StatsStateImpl.kt
@@ -3,12 +3,16 @@ package com.algolia.instantsearch.compose.stats.internal
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import com.algolia.instantsearch.compose.stats.StatsCompose
+import com.algolia.instantsearch.compose.stats.StatsState
 
 /**
- * [StatsCompose] implementation.
+ * [StatsState] implementation.
  */
-internal class StatsComposeImpl<T>(stats: T) : StatsCompose<T> {
+internal class StatsStateImpl<T>(stats: T) : StatsState<T> {
 
     override var stats by mutableStateOf(stats)
+
+    override fun setText(text: T) {
+        this.stats = text
+    }
 }
